fix(wire): figures the venue can state that this client could not carry
A caller's contract number was taken in full and sent narrowed, so an id above
what the request carries went out as a different contract entirely — registered
here under the number the caller gave, and answered with another instrument's
prices and depth. It is refused now, naming the id, rather than quietly asking
about something else.

A count of trades on a bar was held wider than the width every surface reports
it in, so a bar made by more trades than that width carries reached the caller
as a bar made by minus two billion. A count that will not fit is not one this
client can carry, and the bar states none.

An acknowledgement states the increment every price and size on the instrument
is counted in. Zero counts nothing, a negative one turns every figure upside
down, and an infinity — which the parser reads from the word — saturates them
all. None of the three is a scale, and the acknowledgement is refused with the
request named rather than taken as one.

And the arithmetic that follows an execution added and subtracted quantities
without checking, so a report stating a figure large enough could wrap what the
order had filled and what the account held.

+++ src/api/client/dispatch.rs
@@ -566,7 +566,7 @@ impl EClient {
                     close: bar.close,
                     volume: bar.volume,
                     wap: bar.wap,
-                    bar_count: bar.count as i32,
+                    bar_count: bar.count,
                     timezone: response.timezone.clone(),
                 };
                 if is_update {
